Added constructor to HTTP transports which takes a Azure::Core::Http::Policies::TransportOptions
