+++ sdk/core/azure-core/inc/azure/core/http/policies/policy.hpp
@@ -137,7 +137,6 @@ namespace Azure { namespace Core { namespace Http { namespace Policies {
    */
   struct TransportOptions final
   {
-#if !defined(BUILD_TRANSPORT_CUSTOM_ADAPTER)
     /**
      * @brief The URL for the proxy server to use for this connection.
      *
@@ -172,7 +171,6 @@ namespace Azure { namespace Core { namespace Http { namespace Policies {
      * This means that the expected root certificate is the only certificate that will be trusted.
      */
     std::string ExpectedTlsRootCertificate{};
-#endif // !defined(BUILD_TRANSPORT_CUSTOM_ADAPTER)
 
     /**
      * @brief #Azure::Core::Http::HttpTransport that the transport policy will use to send and
