+++ sdk/core/azure-core/test/ut/azure_core_test.cpp
@@ -5,8 +5,13 @@
 
 #if defined(BUILD_CURL_HTTP_TRANSPORT_ADAPTER)
 #include <azure/core/platform.hpp>
-#include <signal.h>
 #endif
+#include <iostream>
+#include <signal.h>
+
+// C Runtime Library errors trigger SIGABRT. Catch that signal handler and report it on the test
+// log.
+void AbortHandler(int signal) { std::cout << "abort() has been called: " << signal << std::endl; }
 
 int main(int argc, char** argv)
 {
@@ -15,6 +20,8 @@ int main(int argc, char** argv)
   // End users need to decide if SIGPIPE should be ignored or not.
   signal(SIGPIPE, SIG_IGN);
 #endif
+
+  signal(SIGABRT, AbortHandler);
 
   testing::InitGoogleTest(&argc, argv);
   auto r = RUN_ALL_TESTS();
